requirements.txt
APScheduler==3.9.1.post1
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
coloredlogs==15.0.1
humanfriendly==10.0
idna==3.4
pyaes==1.6.1
pyasn1==0.4.8
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.4.0
requests==2.28.1
rsa==4.9
schedule==1.1.0
six==1.16.0
Telethon==1.26.1
tzdata==2022.7
tzlocal==4.2
ujson==5.6.0
urllib3==1.26.13
